fix no-op cast
This cast was doing nothing as n was already and int. Looking at math_logstar_n
above, it seems clear this was meant to be casting to a double.

+++ lib/ortho/trapezoid.c
@@ -1029,7 +1029,7 @@ static int math_N(int n, int h)
   int i;
   double v;
 
-  for (i = 0, v = (int) n; i < h; i++)
+  for (i = 0, v = (double) n; i < h; i++)
       v = log2(v);
 
   return (int) ceil((double) 1.0*n/v);
